requirements.txt
google.cloud.storage
google.cloud.oath2